+++ examples/demo/follow.rs
@@ -96,7 +96,7 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
                 .insert_tween_here(
                     Duration::from_secs(2),
                     EaseKind::CubicInOut,
-                    jeb.with_closure(|transform: &mut Transform, value| {
+                    jeb.with_closure(|transform: &mut Transform, value, _| {
                         let start = 0.;
                         let end = TAU;
                         transform.rotation =