+++ src/interpolate.rs
@@ -11,6 +11,7 @@
 //! - [`AngleZ`]
 //! - [`SpriteColor`]
 //! - [`ColorMaterial`]
+//! - All their delta variants (such as [`TranslationDelta`])
 //!
 //! # Your own [`Interpolator`]
 //!
@@ -45,7 +46,7 @@
 //!     type Item = Foo;
 //!
 //!     // Then we define how we want to interpolate `Foo`
-//!     fn interpolate(&self, item: &mut Self::Item, value: f32) {
+//!     fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
 //!         // Usually if the type already have the `.lerp` function provided
 //!         // by the `FloatExt` trait then we can use just that
 //!         item.0 = self.start.lerp(self.end, value);
@@ -90,13 +91,13 @@ use bevy::prelude::*;
 /// Alias for an `Interpolator` as a boxed trait object.
 pub type BoxedInterpolator<Item> = Box<dyn Interpolator<Item = Item>>;
 
-type InterpolatorClosure<I> = Box<dyn Fn(&mut I, f32) + Send + Sync + 'static>;
+type InterpolatorClosure<I> = Box<dyn Fn(&mut I, f32, f32) + Send + Sync + 'static>;
 
 /// Create boxed closure in order to be used with dynamic [`Interpolator`]
 pub fn closure<I, F>(f: F) -> InterpolatorClosure<I>
 where
     I: 'static,
-    F: Fn(&mut I, f32) + Send + Sync + 'static,
+    F: Fn(&mut I, f32, f32) + Send + Sync + 'static,
 {
     Box::new(f)
 }
@@ -115,7 +116,7 @@ pub trait Interpolator: Send + Sync + 'static {
     /// The value should be already sampled from an [`Interpolation`]
     ///
     /// [`Interpolation`]: crate::interpolation::Interpolation
-    fn interpolate(&self, item: &mut Self::Item, value: f32);
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32);
 }
 
 // /// Reflect [`Interpolator`] trait
@@ -197,7 +198,7 @@ pub trait Interpolator: Send + Sync + 'static {
 
 /// Default interpolators
 ///
-/// Register type and systems for the following interpolators:
+/// Register type and systems for the following interpolators and their delta interpolators:
 /// - [`Translation`]
 /// - [`Rotation`]
 /// - [`Scale`]
@@ -217,29 +218,47 @@ impl Plugin for DefaultInterpolatorsPlugin {
             tween::component_tween_system::<Rotation>(),
             tween::component_tween_system::<Scale>(),
             tween::component_tween_system::<AngleZ>(),
+            tween::component_tween_system::<TranslationDelta>(),
+            tween::component_tween_system::<RotationDelta>(),
+            tween::component_tween_system::<ScaleDelta>(),
+            tween::component_tween_system::<AngleZDelta>()
         ))
         .register_type::<tween::ComponentTween<Translation>>()
         .register_type::<tween::ComponentTween<Rotation>>()
         .register_type::<tween::ComponentTween<Scale>>()
-        .register_type::<tween::ComponentTween<AngleZ>>();
+        .register_type::<tween::ComponentTween<AngleZ>>()
+        .register_type::<tween::ComponentTween<TranslationDelta>>()
+        .register_type::<tween::ComponentTween<RotationDelta>>()
+        .register_type::<tween::ComponentTween<ScaleDelta>>()
+        .register_type::<tween::ComponentTween<AngleZDelta>>();
 
         #[cfg(feature = "bevy_sprite")]
-        app.add_tween_systems(tween::component_tween_system::<SpriteColor>())
-            .register_type::<tween::ComponentTween<SpriteColor>>();
+        app.add_tween_systems((
+            tween::component_tween_system::<SpriteColor>(),
+            tween::component_tween_system::<SpriteColorDelta>(),
+        ))
+            .register_type::<tween::ComponentTween<SpriteColor>>()
+            .register_type::<tween::ComponentTween<SpriteColorDelta>>();
 
         #[cfg(feature = "bevy_ui")]
         app.add_tween_systems((
             tween::component_tween_system::<ui::BackgroundColor>(),
             tween::component_tween_system::<ui::BorderColor>(),
+            tween::component_tween_system::<BackgroundColorDelta>(),
+            tween::component_tween_system::<BorderColorDelta>(),
         ))
-        .register_type::<tween::ComponentTween<ui::BackgroundColor>>()
-        .register_type::<tween::ComponentTween<ui::BorderColor>>();
+            .register_type::<tween::ComponentTween<ui::BackgroundColor>>()
+            .register_type::<tween::ComponentTween<ui::BorderColor>>()
+            .register_type::<tween::ComponentTween<BackgroundColorDelta>>()
+            .register_type::<tween::ComponentTween<BorderColorDelta>>();
 
         #[cfg(all(feature = "bevy_sprite", feature = "bevy_asset",))]
-        app.add_tween_systems(
+        app.add_tween_systems((
             tween::asset_tween_system::<sprite::ColorMaterial>(),
-        )
-        .register_type::<tween::AssetTween<sprite::ColorMaterial>>();
+            tween::asset_tween_system::<ColorMaterialDelta>(),
+        ))
+            .register_type::<tween::AssetTween<sprite::ColorMaterial>>()
+            .register_type::<tween::ComponentTween<ColorMaterialDelta>>();
     }
 }