+++ src/interpolate/blanket_impl.rs
@@ -7,8 +7,8 @@ where
 {
     type Item = I::Item;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        (**self).interpolate(item, value, previous_value)
     }
 }
 
@@ -18,8 +18,8 @@ where
 {
     type Item = I::Item;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        (**self).interpolate(item, value, previous_value)
     }
 }
 
@@ -29,15 +29,15 @@ where
 {
     type Item = I::Item;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        (**self).interpolate(item, value)
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        (**self).interpolate(item, value, previous_value)
     }
 }
 
-impl<I: 'static> Interpolator for dyn Fn(&mut I, f32) + Send + Sync + 'static {
+impl<I: 'static> Interpolator for dyn Fn(&mut I, f32, f32) + Send + Sync + 'static {
     type Item = I;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
-        self(item, value)
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        self(item, value, previous_value)
     }
 }