+++ src/interpolate/sprite.rs
@@ -16,10 +16,33 @@ pub struct SpriteColor {
 impl Interpolator for SpriteColor {
     type Item = Sprite;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.color = self.start.mix(&self.end, value)
     }
 }
+
+
+/// delta [`Interpolator`] for [`Sprite`]'s color
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct SpriteColorDelta {
+    #[allow(missing_docs)]
+    pub start: Color,
+    #[allow(missing_docs)]
+    pub end: Color,
+}
+
+impl Interpolator for SpriteColorDelta {
+    type Item = Sprite;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_color_as_vec = self.start.mix(&self.end, previous_value).to_linear().to_vec4();
+        let next_color_as_vec = self.start.mix(&self.end, value).to_linear().to_vec4();
+        let color_delta = next_color_as_vec - previous_color_as_vec;
+        let updated_color = item.color.to_linear().to_vec4() + color_delta;
+        item.color = Color::srgba(updated_color.x, updated_color.y, updated_color.z, updated_color.w);
+    }
+}
+
 
 /// Constructor for [`SpriteColor`]
 pub fn sprite_color(start: Color, end: Color) -> SpriteColor {
@@ -39,7 +62,7 @@ pub fn sprite_color_to(to: Color) -> impl Fn(&mut Color) -> SpriteColor {
 // type ReflectInterpolatorColorMaterial =
 //     ReflectInterpolator<bevy::sprite::ColorMaterial>;
 
-/// [`Interpolator`] for [`Sprite`]'s [`ColorMaterial`]
+/// [`Interpolator`] for [`Sprite`]'s ColorMaterial
 #[derive(Debug, Default, Clone, PartialEq, Reflect)]
 // #[reflect(InterpolatorColorMaterial)]
 pub struct ColorMaterial {
@@ -52,10 +75,32 @@ pub struct ColorMaterial {
 impl Interpolator for ColorMaterial {
     type Item = bevy::sprite::ColorMaterial;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.color = self.start.mix(&self.end, value);
     }
 }
+
+/// delta [`Interpolator`] for [`Sprite`]'s ColorMaterial
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct ColorMaterialDelta {
+    #[allow(missing_docs)]
+    pub start: Color,
+    #[allow(missing_docs)]
+    pub end: Color,
+}
+
+impl Interpolator for ColorMaterialDelta {
+    type Item = bevy::sprite::ColorMaterial;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_color_as_vec = self.start.mix(&self.end, previous_value).to_linear().to_vec4();
+        let next_color_as_vec = self.start.mix(&self.end, value).to_linear().to_vec4();
+        let color_delta = next_color_as_vec - previous_color_as_vec;
+        let updated_color = item.color.to_linear().to_vec4() + color_delta;
+        item.color = Color::srgba(updated_color.x, updated_color.y, updated_color.z, updated_color.w);
+    }
+}
+
 
 /// Constructor for [`ColorMaterial`](crate::interpolate::ColorMaterial)
 pub fn color_material(start: Color, end: Color) -> ColorMaterial {