+++ src/interpolate/transform.rs
@@ -15,8 +15,27 @@ pub struct Translation {
 impl Interpolator for Translation {
     type Item = Transform;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.translation = self.start.lerp(self.end, value);
+    }
+}
+
+/// delta [`Interpolator`] for [`Transform`]'s translation.
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct TranslationDelta {
+    #[allow(missing_docs)]
+    pub start: Vec3,
+    #[allow(missing_docs)]
+    pub end: Vec3,
+}
+impl Interpolator for TranslationDelta {
+    type Item = Transform;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_translation = self.start.lerp(self.end, previous_value);
+        let next_translation = self.start.lerp(self.end, value);
+        let translation_delta = next_translation - previous_translation;
+        item.translation += translation_delta;
     }
 }
 
@@ -45,6 +64,16 @@ pub fn translation_by(by: Vec3) -> impl Fn(&mut Vec3) -> Translation {
     }
 }
 
+/// Constructor for [`TranslationDelta`] that's relative to previous value
+/// Since this is a delta tween, it can happen with other ongoing tweens of that type
+pub fn translation_delta_by(by: Vec3) -> impl Fn(&mut Vec3) -> TranslationDelta {
+    move |state| {
+        let start = *state;
+        let end = *state + by;
+        TranslationDelta { start, end }
+    }
+}
+
 /// [`Interpolator`] for [`Transform`]'s rotation using the [`Quat::slerp`] function.
 #[derive(Debug, Default, Clone, PartialEq, Reflect)]
 // #[reflect(InterpolatorTransform)]
@@ -57,8 +86,27 @@ pub struct Rotation {
 impl Interpolator for Rotation {
     type Item = Transform;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.rotation = self.start.slerp(self.end, value);
+    }
+}
+
+/// delta [`Interpolator`] for [`Transform`]'s rotation using the [`Quat::slerp`] function.
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct RotationDelta {
+    #[allow(missing_docs)]
+    pub start: Quat,
+    #[allow(missing_docs)]
+    pub end: Quat,
+}
+impl Interpolator for RotationDelta {
+    type Item = Transform;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_rotation = self.start.slerp(self.end, previous_value);
+        let next_rotation = self.start.slerp(self.end, value);
+        let rotation_delta = next_rotation - previous_rotation;
+        item.rotation = item.rotation.mul_quat(rotation_delta);
     }
 }
 
@@ -87,6 +135,18 @@ pub fn rotation_by(by: Quat) -> impl Fn(&mut Quat) -> Rotation {
     }
 }
 
+
+/// Constructor for [`RotationDelta`] that's relative to previous value
+/// Since this is a delta tween, it can happen with other ongoing tweens of that type
+pub fn rotation_delta_by(by: Quat) -> impl Fn(&mut Quat) -> RotationDelta {
+    move |state| {
+        let start = *state;
+        let end = *state + by;
+        *state = state.mul_quat(by);
+        RotationDelta { start, end }
+    }
+}
+
 /// [`Interpolator`] for [`Transform`]'s scale
 #[derive(Debug, Default, Clone, PartialEq, Reflect)]
 // #[reflect(InterpolatorTransform)]
@@ -99,8 +159,28 @@ pub struct Scale {
 impl Interpolator for Scale {
     type Item = Transform;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.scale = self.start.lerp(self.end, value);
+    }
+}
+
+
+/// delta [`Interpolator`] for [`Transform`]'s scale
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct ScaleDelta {
+    #[allow(missing_docs)]
+    pub start: Vec3,
+    #[allow(missing_docs)]
+    pub end: Vec3,
+}
+impl Interpolator for ScaleDelta {
+    type Item = Transform;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_scale = self.start.lerp(self.end, previous_value);
+        let next_scale = self.start.lerp(self.end, value);
+        let scale_delta = next_scale - previous_scale;
+        item.scale += scale_delta;
     }
 }
 
@@ -129,6 +209,17 @@ pub fn scale_by(by: Vec3) -> impl Fn(&mut Vec3) -> Scale {
     }
 }
 
+/// Constructor for [`ScaleDelta`] that's relative to previous value
+/// Since this is a delta tween, it can happen with other ongoing tweens of that type
+pub fn scale_delta_by(by: Vec3) -> impl Fn(&mut Vec3) -> ScaleDelta {
+    move |state| {
+        let start = *state;
+        let end = *state + by;
+        *state += by;
+        ScaleDelta { start, end }
+    }
+}
+
 /// [`Interpolator`] for [`Transform`]'s rotation at Z axis.
 /// Usually used for 2D rotation.
 #[derive(Debug, Default, Clone, PartialEq, Reflect)]
@@ -142,9 +233,29 @@ pub struct AngleZ {
 impl Interpolator for AngleZ {
     type Item = Transform;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         let angle = (self.end - self.start).mul_add(value, self.start);
         item.rotation = Quat::from_rotation_z(angle);
+    }
+}
+
+/// [`Interpolator`] for [`Transform`]'s rotation at Z axis.
+/// Usually used for 2D rotation.
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct AngleZDelta {
+    #[allow(missing_docs)]
+    pub start: f32,
+    #[allow(missing_docs)]
+    pub end: f32,
+}
+impl Interpolator for AngleZDelta {
+    type Item = Transform;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_angle = (self.end - self.start).mul_add(previous_value, self.start);
+        let update_angle = (self.end - self.start).mul_add(value, self.start);
+        let angle_delta_as_quat = Quat::from_rotation_z(update_angle - previous_angle);
+        item.rotation = item.rotation.mul_quat(angle_delta_as_quat);
     }
 }
 
@@ -170,5 +281,16 @@ pub fn angle_z_by(by: f32) -> impl Fn(&mut f32) -> AngleZ {
         let end = *state + by;
         *state += by;
         angle_z(start, end)
+    }
+}
+
+/// Constructor for [`AngleZDelta`] that's relative to previous value
+/// Since this is a delta tween, it can happen with other ongoing tweens of that type
+pub fn angle_z_delta_by(by: f32) -> impl Fn(&mut f32) -> AngleZDelta {
+    move |state| {
+        let start = *state;
+        let end = *state + by;
+        *state += by;
+        AngleZDelta {start, end}
     }
 }