+++ src/interpolate/ui.rs
@@ -13,8 +13,29 @@ pub struct BackgroundColor {
 impl Interpolator for BackgroundColor {
     type Item = bevy::prelude::BackgroundColor;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.0 = self.start.mix(&self.end, value)
+    }
+}
+
+/// delta [`Interpolator`] for Bevy's [`BackgroundColor`](bevy::prelude::BackgroundColor) used in UIs.
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct BackgroundColorDelta {
+    #[allow(missing_docs)]
+    pub start: Color,
+    #[allow(missing_docs)]
+    pub end: Color,
+}
+
+impl Interpolator for BackgroundColorDelta {
+    type Item = bevy::prelude::BackgroundColor;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_color_as_vec = self.start.mix(&self.end, previous_value).to_linear().to_vec4();
+        let next_color_as_vec = self.start.mix(&self.end, value).to_linear().to_vec4();
+        let color_delta = next_color_as_vec - previous_color_as_vec;
+        let updated_color = item.0.to_linear().to_vec4() + color_delta;
+        item.0 = Color::srgba(updated_color.x, updated_color.y, updated_color.z, updated_color.w);
     }
 }
 
@@ -47,10 +68,33 @@ pub struct BorderColor {
 impl Interpolator for BorderColor {
     type Item = bevy::prelude::BorderColor;
 
-    fn interpolate(&self, item: &mut Self::Item, value: f32) {
+    fn interpolate(&self, item: &mut Self::Item, value: f32, _previous_value: f32) {
         item.0 = self.start.mix(&self.end, value)
     }
 }
+
+
+/// delta [`Interpolator`] for Bevy's [`BorderColor`](bevy::prelude::BorderColor) used in UIs.
+#[derive(Debug, Default, Clone, PartialEq, Reflect)]
+pub struct BorderColorDelta {
+    #[allow(missing_docs)]
+    pub start: Color,
+    #[allow(missing_docs)]
+    pub end: Color,
+}
+
+impl Interpolator for BorderColorDelta {
+    type Item = bevy::prelude::BackgroundColor;
+
+    fn interpolate(&self, item: &mut Self::Item, value: f32, previous_value: f32) {
+        let previous_color_as_vec = self.start.mix(&self.end, previous_value).to_linear().to_vec4();
+        let next_color_as_vec = self.start.mix(&self.end, value).to_linear().to_vec4();
+        let color_delta = next_color_as_vec - previous_color_as_vec;
+        let updated_color = item.0.to_linear().to_vec4() + color_delta;
+        item.0 = Color::srgba(updated_color.x, updated_color.y, updated_color.z, updated_color.w);
+    }
+}
+
 
 /// Constructor for [`BorderColor`](crate::interpolate::BorderColor)
 pub fn border_color(start: Color, end: Color) -> BorderColor {